+++ StenGine/GLImageLoader.h
@@ -2,8 +2,9 @@
 #define __GL_IMAGE_LOADER_H_
 
 #include "GL/glew.h"
+#include <stdint.h>
 
-GLuint CreateGLTextureFromFile(const char* filename);
+uint64_t CreateGLTextureFromFile(const char* filename);
 
 
 #endif
